update goblin from 0.9 to 0.10, parse PEs permissively
goblin 0.10's ParseMode::Permissive skips the DOS stub validation that
rejected PEs whose header overlaps the DOS header (e_lfanew < 0x40), which
the Windows loader accepts. opt into permissive parsing in get_pe so such
files (e.g. the tiny.exe test resource) load again.

restore loader::pe::tests::tiny from an expected-failure test back to
asserting the module loads, re-add the tiny rows to the analysis test
matrices, and keep the cpp1 rows added alongside.

only API change needed: address_of_entry_point is now u32.

+++ core/src/emu/plat/win/win32.rs
@@ -180,7 +180,7 @@ mod tests {
         emu.load_pe(&pe)?;
 
         let opt = pe.optional_header.unwrap();
-        let ep = opt.windows_fields.image_base + opt.standard_fields.address_of_entry_point;
+        let ep = opt.windows_fields.image_base + opt.standard_fields.address_of_entry_point as u64;
         emu.set_pc(ep);
 
         emu.mem().mmap(0x5000, 0x2000, Permissions::RW)?;
